+++ src/lib.rs
@@ -244,7 +244,8 @@ pub use api::{
     BracketType, EasyToBorrowRequest, LoginConfig, ModifyPriceType, NewOrderDuration,
     NewOrderPriceType, NewOrderTransactionType, OcoDuration, OcoPriceType, OcoTransactionType,
     RithmicAdvancedBracketOrder, RithmicBracketOrder, RithmicCancelOrder, RithmicIfTouchedTrigger,
-    RithmicModifyOrder, RithmicOcoOrderLeg, RithmicOrder, RithmicResponse, TrailingStop,
+    RithmicModifyOrder, RithmicOcoOrderLeg, RithmicOrder, RithmicResponse, RmsUpdateBits,
+    TrailingStop,
 };
 
 // Re-export utility types for convenience